pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ievo-curator"
version = "0.1.0"
description = "Level 2 evolution — cross-agent pattern detection and shared skill updates for iEvo marketplace"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "iEvo", email = "denis@27tech.co" }]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]

[project.scripts]
curator = "curator.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/curator"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
